minor fixes & improvements for lists resolvers

+++ src/api/List/deleteList/resolvers.ts
@@ -2,7 +2,10 @@ import {Resolvers} from '@/types/generated'
 
 const resolvers: Resolvers = {
   Mutation: {
-    deleteList: async (_, {id}, {prisma}) => {
+    deleteList: async (_, {id}, {prisma, user}) => {
+      if (!user)
+        return {ok: false, error: 'Unauthorized Access: user not logged in'}
+
       try {
         await prisma.list.delete({
           where: {id},
